LoginPage ui - fixed suggestions1
