+++ lib/views/screens/login_page.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:sizer/sizer.dart';
+
+class LoginPage extends StatefulWidget {
+  const LoginPage({super.key});
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: Scaffold(
+        backgroundColor: Colors.white,
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          children: [
+            Container(
+              alignment: Alignment.center,
+              child: Image.asset(
+                "assets/images/Memojiswelcome.jpg",
+              ),
+            ),
+            Text(
+              "Find your friends.\nAnytime. Anywhere.",
+              style: GoogleFonts.inter(
+                textStyle: Theme.of(context)
+                    .textTheme
+                    .headlineMedium!
+                    .copyWith(fontWeight: FontWeight.bold),
+              ),
+            ),
+            FilledButton(
+              onPressed: () {},
+              style: ButtonStyle(
+                fixedSize: WidgetStatePropertyAll(
+                  Size(80.w, 7.h),
+                ),
+                shape: WidgetStatePropertyAll(
+                  RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(14),
+                  ),
+                ),
+              ),
+              child: const Text("SignIn With Outlook"),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
