+++ pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons:
   google_nav_bar:
+  sizer:
+  google_fonts: 
 
 dev_dependencies:
   flutter_test:
